+++ src/features/XIT/CXTS/CXTS.vue
@@ -18,7 +18,7 @@ interface OrderTrade {
 interface DayTrades {
   date: number;
   trades: OrderTrade[];
-  totals: { [currency: string]: number };
+  totals: { [currency: string]: { purchases: number; sales: number } };
 }
 
 const msInDay = dayjs.duration(1, 'day').asMilliseconds();
@@ -63,10 +63,14 @@ const days = computed(() => {
     }
 
     day.trades.push(trade);
-    const total =
-      trade.trade.price.amount * trade.trade.amount * (trade.order.type === 'SELLING' ? 1 : -1);
     const currency = trade.trade.price.currency;
-    day.totals[currency] = (day.totals[currency] ?? 0) + total;
+    const total = trade.trade.price.amount * trade.trade.amount;
+    const totals = (day.totals[currency] ??= { purchases: 0, sales: 0 });
+    if (trade.order.type === 'SELLING') {
+      totals.sales += total;
+    } else {
+      totals.purchases += total;
+    }
   }
   return days;
 });